extract wordsFilter interface
