+++ cs/Homework/TagsCloud.Visualization/WordsFilter/IWordsFilter.cs
@@ -1,0 +1,7 @@
+﻿namespace TagsCloud.Visualization.WordsFilter
+{
+    public interface IWordsFilter
+    {
+        bool IsWordValid(string word);
+    }
+}
